Initial commit: HL7 parser works.
This HL7 parser creates a parse tree of Clojure records.  Escape
character resolution is not implemented yet, and neither is message
generation.  Hopefully this is the start of a very nice competitor to
Cloverleaf and Mirth.  This draws on some work by Christopher Miles
(@cmiles74).

Repeating fields and subcomponents work properly.